pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ghibli"
version = "0.1.0"
description = "GitHub Intelligence Bridge — natural language to GitHub REST API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "google-genai>=1.0",
    "rich>=13",
    "httpx>=0.27",
    "python-dotenv>=1",
    "pyyaml>=6.0.3",
    "litellm>=1.50.0",
]

[project.scripts]
ghibli = "ghibli.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "ruff>=0.4",
    "black>=24",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--cov=src/ghibli --cov-report=term-missing --cov-fail-under=80"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.black]
line-length = 88

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
]
